combined DefaultDiktatRuleSetFactory and DiktatRuleSetFactory
