+++ diktat-rules/src/main/kotlin/org/cqfn/diktat/ruleset/rules/DiktatRuleSetProvider.kt
@@ -1,6 +1,6 @@
 package org.cqfn.diktat.ruleset.rules
 
-import org.cqfn.diktat.ktlint.KtLintRuleSetWrapper
+import org.cqfn.diktat.ktlint.KtLintRuleSetWrapper.Companion.toKtLint
 import com.pinterest.ktlint.core.RuleSet
 import com.pinterest.ktlint.core.RuleSetProvider
 
@@ -9,9 +9,8 @@ import com.pinterest.ktlint.core.RuleSetProvider
  * By default, it is expected to have `diktat-analysis.yml` configuration in the root folder where 'ktlint' is run
  * otherwise it will use default configuration where some rules are disabled.
  *
- * @param diktatRuleSetFactory  the configuration file where all configurations for
- *   inspections and rules are stored.
+ * This class is registered in [resources/META-INF/services/com.pinterest.ktlint.core.RuleSetProvider]
  */
-class DiktatRuleSetProvider(private val diktatRuleSetFactory: DiktatRuleSetFactory = DiktatRuleSetFactory()) : RuleSetProvider {
-    override fun get(): RuleSet = KtLintRuleSetWrapper(diktatRuleSetFactory())
+class DiktatRuleSetProvider : RuleSetProvider {
+    override fun get(): RuleSet = DiktatRuleSetFactory().invoke().toKtLint()
 }
