+++ diktat-rules/src/test/kotlin/org/cqfn/diktat/ruleset/smoke/RulesConfigValidationTest.kt
@@ -1,7 +1,7 @@
 package org.cqfn.diktat.ruleset.smoke
 
+import org.cqfn.diktat.ktlint.KtLintRuleSetProvider.Companion.toKtLint
 import org.cqfn.diktat.ruleset.rules.DiktatRuleSetFactory
-import org.cqfn.diktat.ruleset.rules.DiktatRuleSetProvider
 import org.cqfn.diktat.test.framework.util.deleteIfExistsSilently
 
 import com.charleskorn.kaml.InvalidPropertyValueException
@@ -40,8 +40,7 @@ class RulesConfigValidationTest {
             """.trimMargin()
         )
         val exception = assertThrows<IllegalArgumentException> {
-            @Suppress("DEPRECATION")
-            DiktatRuleSetProvider(DiktatRuleSetFactory(file.absolutePath)).get()
+            DiktatRuleSetFactory(file.absolutePath).toKtLint().get()
         }
         Assertions.assertEquals("Warning name <MISSING_DOC_TOP_LEVEL> in configuration file is invalid, did you mean <MISSING_KDOC_TOP_LEVEL>?", exception.message)
     }
@@ -56,8 +55,7 @@ class RulesConfigValidationTest {
             """.trimMargin()
         )
         assertThrows<InvalidPropertyValueException> {
-            @Suppress("DEPRECATION")
-            DiktatRuleSetProvider(DiktatRuleSetFactory(file.absolutePath)).get()
+            DiktatRuleSetFactory(file.absolutePath).toKtLint().get()
         }
     }
 
@@ -73,7 +71,6 @@ class RulesConfigValidationTest {
                 |    isIncludeHeader: Fslse
             """.trimMargin()
         )
-        @Suppress("DEPRECATION")
-        DiktatRuleSetProvider(DiktatRuleSetFactory(file.absolutePath)).get()
+        DiktatRuleSetFactory(file.absolutePath).toKtLint().get()
     }
 }
