+++ diktat-rules/src/test/kotlin/org/cqfn/diktat/util/DiktatRuleSetProvider4Test.kt
@@ -8,14 +8,13 @@
 
 package org.cqfn.diktat.util
 
-import org.cqfn.diktat.common.config.rules.DIKTAT_RULE_SET_ID
 import org.cqfn.diktat.common.config.rules.RulesConfig
 import org.cqfn.diktat.common.config.rules.RulesConfigReader
-import org.cqfn.diktat.ktlint.KtLintRuleSetWrapper
+import org.cqfn.diktat.ktlint.KtLintRuleSetProvider.Companion.toKtLint
 import org.cqfn.diktat.ktlint.KtLintRuleSetWrapper.Companion.delegatee
+import org.cqfn.diktat.ktlint.KtLintRuleSetWrapper.Companion.toKtLint
 import org.cqfn.diktat.ruleset.rules.DiktatRule
 import org.cqfn.diktat.ruleset.rules.DiktatRuleSet
-import org.cqfn.diktat.ruleset.rules.DiktatRuleSetProvider
 import org.cqfn.diktat.test.framework.util.filterContentMatches
 
 import com.pinterest.ktlint.core.Rule
@@ -40,7 +39,7 @@ class DiktatRuleSetProvider4Test(private val ruleSupplier: (rulesConfigList: Lis
     private val rulesConfigList: List<RulesConfig>? = rulesConfigList ?: RulesConfigReader(javaClass.classLoader).readResource("diktat-analysis.yml")
 
     @Suppress("OVERRIDE_DEPRECATION")
-    override fun get(): RuleSet = KtLintRuleSetWrapper(DiktatRuleSet(listOf(ruleSupplier.invoke(rulesConfigList ?: emptyList()))))
+    override fun get(): RuleSet = DiktatRuleSet(listOf(ruleSupplier.invoke(rulesConfigList ?: emptyList()))).toKtLint()
 }
 
 class DiktatRuleSetProviderTest {
@@ -56,7 +55,9 @@ class DiktatRuleSetProviderTest {
             .map(Path::nameWithoutExtension)
             .filterNot { it in ignoredFileNames }
             .toList()
-        val ruleNames = DiktatRuleSetProvider().get()
+        val ruleNames = DiktatRuleSetFactory()
+            .toKtLint()
+            .get()
             .asSequence()
             .onEachIndexed { index, rule ->
                 if (index != 0) {
